+++ NEWS.md
@@ -1,5 +1,6 @@
 # hydflood 0.5.8
 
+* replace scripted html-postprocessing with packaged (`bfgdown::cleanAll`)
 * deactivate time-consuming classifyToPNV() example
 * improve website accessibility
 * move file downloads from `utils::download.file` to `httr2`